// frontend/src/api/generated/model/ndmsVpnServerKind.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * keen-pbr API
 * REST API for the keen-pbr policy-based routing daemon.
 * OpenAPI spec version: 3.0.0
 */

export type NdmsVpnServerKind = typeof NdmsVpnServerKind[keyof typeof NdmsVpnServerKind];


export const NdmsVpnServerKind = {
  l2tp: 'l2tp',
  ikev1: 'ikev1',
  ikev2: 'ikev2',
  sstp: 'sstp',
  openconnect: 'openconnect',
} as const;